+++ logger/config.js
@@ -24,10 +24,11 @@ const DEFAULT_OPTIONS = {
   CONSOLE_LOG_CONFIG: {
     IS_ACTIVE: true,
     LOG_LEVEL: 'debug',
+    IS_JSON : false,
   },
   FILE_LOG_CONFIG: [
-    { FILE_NAME: 'error', LOG_LEVEL: 'error', DATE_PATTERN: 'YYYY-MM-DD-HH' },
-    { FILE_NAME: 'all', LOG_LEVEL: 'debug', DATE_PATTERN: 'YYYY-MM-DD' },
+    { FILE_NAME: 'error', LOG_LEVEL: 'error', DATE_PATTERN: 'YYYY-MM-DD', IS_JSON: true },
+    { FILE_NAME: 'all', LOG_LEVEL: 'debug', DATE_PATTERN: 'YYYY-MM-DD', IS_JSON: true },
   ],
 
   BANNER_STRING: `logger pord by manbalboy