+++ CHANGELOG.md
@@ -1,4 +1,21 @@
 # Changelog
+## [0.0.5] 2022-01-24
+### Bug fix
+- utils error fix
+
+## [0.0.4] 2022-01-24
+### Bug fix
+- Converting Circular Structure to JSON error handling during JSON parsing
+### add function
+- Change the log display method to be set in each logger setting
+- File log defaults to JSON, Console log defaults to not using JSON.
+
+## [0.0.3] 2022-01-24
+### Bug fix
+- default variable setting
+
+## [0.0.2] 2022-01-23
+### README.md file update
 
 ## [0.0.1] 2022-01-23
 ### Stable Original Release